Add files via upload
In this file, I am learning about arrays in JavaScript and Data Structuring with Set.
Finally, the data structure we are discussing is set. Set is a special data structure compared to maps, arrays, and objects. Why is it called special?

The answer is because a set doesn't have keys and indices when storing data. Additionally, data stored in a set will be unique, meaning there won't be any duplicate data.