Add files via upload
In this file, I am learning about arrays in JavaScript - specifically about the Spread Operator.
Spread, true to its name, means to distribute or scatter. The spread operator is used to spread out values that exist in objects and arrays. The spread operator, marked by three dots (...), is an interesting feature that helps in managing objects and arrays. By using the spread operator, object and array values can be iterated into multiple elements.