Add files via upload
In this JavaScript file, there is material about GET & SET, Getter and Setter: Functions to read and modify property values with specific control. A Getter consists of the get method, which is a way to retrieve the value of a property, while a Setter consists of the set method, which is used to assign a value to a property. This allows us to manage access to the properties owned by an object.