Add files via upload
Object Composition is a design technique in programming that allows the creation of complex objects by combining various simpler objects or functions. It serves as an alternative to inheritance in object-oriented programming.

Object Composition is more flexible than inheritance because it does not require objects to adhere to a rigid hierarchy. Instead, it enables code reuse by combining features from multiple sources.