Add files via upload
A pure function refers to a function that adheres to the following two properties:

1.Produces the same output for the same input every time it is called.
For example, when a function is given the argument X, it must always return the value Y, regardless of the conditions or the time of execution. This ensures that the function’s output is predictable. To achieve this, the function must not access values outside of its arguments or rely on global variables.

2.Has no side effects that impact the state outside the function.
In this context, side effects refer to any changes made by the function to variables or states outside its scope, such as modifying global variables or interacting with input and output operations.